SwiftUI - 21. Project Structure and AppStorage
