+++ Swift/SwiftUI for iOS 15/DesignCodeiOS15a/DesignCodeiOS15a/DesignCodeiOS15aApp.swift
@@ -11,7 +11,7 @@ import SwiftUI
 struct DesignCodeiOS15aApp: App {
 	var body: some Scene {
 		WindowGroup {
-			TabBar()
+			ContentView()
 		}
 	}
 }
