+++ Swift/SwiftUI for iOS 15/DesignCodeiOS15a/DesignCodeiOS15a/Navigation/TabBar.swift
@@ -9,45 +9,27 @@ import SwiftUI
 
 struct TabBar: View {
 	
-	@State var selectedTab: Tab = .home
+	@AppStorage("selectedTab") var selectedTab: Tab = .home
+	
 	@State var selectedTabColor: Color = .teal
 	@State var tabItemWidth: CGFloat = 0
 	
     var body: some View {
-		
-		ZStack(alignment: .bottom) {
-			
-			Group {
-				switch selectedTab {
-					case .home:
-						ContentView()
-					case .explore:
-						AccountView()
-					case .notifications:
-						AccountView()
-					case .library:
-						AccountView()
-				}
-			}
-			.frame(maxWidth: .infinity, maxHeight: .infinity)
-			.ignoresSafeArea()
-			
-			HStack {
-				buttons
-			}
-			.padding(.horizontal, 8)
-			.padding(.top, 14)
-			.frame(height: 88, alignment: .top)
-			.background(.ultraThinMaterial, in: RoundedRectangle(cornerRadius: 34, style: .continuous))
-			.background(
-				background
-			)
-			.overlay(
-				overlay
-			)
-			.strokeStyle(cornerRadius: 34)
-			.frame(maxHeight: .infinity, alignment: .bottom)
+		HStack {
+			buttons
 		}
+		.padding(.horizontal, 8)
+		.padding(.top, 14)
+		.frame(height: 88, alignment: .top)
+		.background(.ultraThinMaterial, in: RoundedRectangle(cornerRadius: 34, style: .continuous))
+		.background(
+			background
+		)
+		.overlay(
+			overlay
+		)
+		.strokeStyle(cornerRadius: 34)
+		.frame(maxHeight: .infinity, alignment: .bottom)
     }
 	
 	var buttons: some View {
