+++ Swift/SwiftUI for iOS 15/DesignCodeiOS15a/DesignCodeiOS15a/Views/AccountView.swift
@@ -56,13 +56,13 @@ struct AccountView: View {
 	
 	var menu: some View {
 		Section {
-			NavigationLink(destination: ContentView()) {
+			NavigationLink(destination: HomeView()) {
 				Label("Settings", systemImage: "gear")
 			}
 			NavigationLink { Text("Billing") } label: {
 				Label("Billing", systemImage: "creditcard")
 			}
-			NavigationLink { ContentView() } label: {
+			NavigationLink { HomeView() } label: {
 				Label("Help", systemImage: "questionmark")
 					.imageScale(.small)
 			}
